finish mapping most of daijirin
Now you can search for totally useful every day words like 瘟㾮日
and 多羅吒干𤚥 :^).

The characters that remain either don't exist in unicode or are very
difficult to find. Also a couple terms seem unsearchable in qolibri so
I couldn't check what the characters are supposed to be.

Any questionable choice was marked with FIXME. This will make it easy in
the future to replace some characters with their images if its something
that we want to support in the future.

* The FIXMEs with the missing font symbol should all be the correct
  character (not commonly covered by fonts)

* The くの字点 choices are to try and imitate the daijirin
  experience(TM). Probably the worst use of image fonts I've seen. Those
  characters should never appear in horizontal text. They should have
  just been replaced with the text that was supposed to be repeated.

* The 漢文訓読 characters in '{}' are technically the unicode specified
  characters for those glyphs however they just look like their full
  size variants. I surrounded them with '{}' so the examples that use
  them are still readable.

* The other FIXMEs should be self explanatory. Search the term in qolibri
  and look at what they used to see why they are questionable.